requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
httpx>=0.27.0
google-generativeai>=0.8.0
gspread>=6.0.0
google-auth>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
twilio>=9.0.0
tzdata>=2024.1
